// src/shared/enums/evaluator-enums.ts
export enum EvaluatorProfile {
  JURIDICO = 'JURIDICO',
  SOCIEDAD_CIVIL = 'SOCIEDAD_CIVIL',
  METODOLOGICO = 'METODOLOGICO',
  SALUD = 'SALUD',
}

export enum AssignmentStatus {
  ASSIGNED = 'ASSIGNED',
  SUBMITTED = 'SUBMITTED',
  REASIGNED_VENCIMIENTO = 'REASIGNED_VENCIMIENTO',
  REASIGNED_COI = 'REASIGNED_COI',
}

export enum ReassignmentReason {
  VENCIMIENTO = 'VENCIMIENTO',
  CONFLICTO_INTERES = 'CONFLICTO_INTERES',
}